Update on Tue Feb 2nd 2021 - Added mySQLConnector.jar

+++ solutions/pluginSolution/source_gen/pluginSolution/plugin/ServiceGenerator.java
@@ -438,6 +438,9 @@ public class ServiceGenerator {
     return conn.createStatement().executeQuery("SHOW DATABASES LIKE '" + databaseName + "'").next();
   }
   private boolean deleteDirectory(File d) {
+    if (!(d.exists())) {
+      return false;
+    }
     File[] files = d.listFiles();
     for (int i = 0; i < files.length; i++) {
       files[i].delete();
